software/requirements.txt
pycryptodome==3.20.0
numpy==2.0.0
scipy==1.14.0
